update keccak test
