+++ x/evm/types/state_object_test.go
@@ -2,9 +2,6 @@ package types_test
 
 import (
 	"math/big"
-	"testing"
-
-	"github.com/okex/exchain/x/evm/types"
 
 	ethcmn "github.com/ethereum/go-ethereum/common"
 )
@@ -135,16 +132,4 @@ func (suite *StateDBTestSuite) TestStateObject_Code() {
 		code := suite.stateObject.Code(nil)
 		suite.Require().Equal(tc.expCode, code, tc.name)
 	}
-}
-
-func TestKeccak256HashWithSyncPool(t *testing.T) {
-	types.TestKeccak256HashWithSyncPool(t)
-}
-
-func BenchmarkKeccak256HashCache(b *testing.B) {
-	types.BenchmarkKeccak256HashCache(b)
-}
-
-func BenchmarkKeccak256HashNew(b *testing.B) {
-	types.BenchmarkKeccak256HashNew(b)
 }
